# server/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Nianxiang API
  version: "1"
  description: |
    Self-hosted memory diary API.
    Auth via Bearer JWT. Entries are strictly owner-isolated; people are household-shared.
servers:
  - url: /api/v1
paths:
...
  /me/profile:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Profile + memories
    patch:
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Updated profile
  /me/memories/{memId}:
    parameters:
      - in: path
        name: memId
        required: true
        schema: { type: string }
    patch:
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Memory edited }
    delete:
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Memory deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []